+++ src/main/java/darkoverload/itzip/feature/techinfo/service/blog/impl/BlogServiceImpl.java
@@ -6,26 +6,28 @@ import darkoverload.itzip.feature.techinfo.controller.response.BlogBasicInfoResp
 import darkoverload.itzip.feature.techinfo.controller.response.BlogDetailInfoResponse;
 import darkoverload.itzip.feature.techinfo.domain.Blog;
 import darkoverload.itzip.feature.techinfo.dto.year.YearlyPostDto;
+import darkoverload.itzip.feature.techinfo.model.entity.BlogEntity;
 import darkoverload.itzip.feature.techinfo.repository.blog.BlogRepository;
 import darkoverload.itzip.feature.techinfo.service.blog.BlogService;
 import darkoverload.itzip.feature.techinfo.service.post.PostService;
-import darkoverload.itzip.feature.techinfo.service.shared.SharedService;
 import darkoverload.itzip.feature.user.domain.User;
+import darkoverload.itzip.feature.user.entity.UserEntity;
+import darkoverload.itzip.feature.user.repository.UserRepository;
+import darkoverload.itzip.global.config.response.code.CommonExceptionCode;
+import darkoverload.itzip.global.config.response.exception.RestApiException;
+import org.springframework.transaction.annotation.Transactional;
+import org.springframework.stereotype.Service;
 
 import lombok.RequiredArgsConstructor;
 
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-
 import java.util.List;
-import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
 public class BlogServiceImpl implements BlogService {
 
     private final BlogRepository blogRepository;
-    private final SharedService sharedService;
+    private final UserRepository userRepository;
     private final PostService postService;
 
     @Override
@@ -38,9 +40,14 @@ public class BlogServiceImpl implements BlogService {
     @Override
     @Transactional
     public void updateBlog(CustomUserDetails userDetails, BlogUpdateRequest request) {
-        Long userId = sharedService.getUserByEmail(userDetails.getEmail()).getId();
-        Blog blog = sharedService.getBlogById(userId);
+        Long userId = userRepository.findByEmail(userDetails.getEmail())
+                .map(UserEntity::convertToDomain)
+                .orElseThrow(
+                        () -> new RestApiException(CommonExceptionCode.NOT_FOUND_USER)
+                )
+                .getId();
 
+        Blog blog = getBlogById(userId);
         blog.setIntro(request.getIntro()); // 블로그 소개글 업데이트
         blogRepository.save(blog.convertToEntity());
     }
@@ -48,27 +55,44 @@ public class BlogServiceImpl implements BlogService {
     @Override
     @Transactional
     public void setBlogHiddenStatus(Long blogId) {
-        Blog blog = sharedService.getBlogById(blogId);
+        Blog blog = getBlogById(blogId);
 
         blog.setIsPublic(false); // 블로그 비공개 설정
         blogRepository.save(blog.convertToEntity());
     }
 
     @Override
+    @Transactional(readOnly = true)
+    public Blog getBlogById(Long id) {
+        return blogRepository.findByIdAndIsPublic(id)
+                .map(BlogEntity::convertToDomain)
+                .orElseThrow(
+                        () -> new RestApiException(CommonExceptionCode.NOT_FOUND_BLOG)
+                );
+    }
+
+    @Override
     public BlogBasicInfoResponse getBasicBlogInfo(Long blogId) {
-        Blog blog = sharedService.getBlogById(blogId);
+        Blog blog = getBlogById(blogId);
         return blog.convertToBlogBasicInfoResponse(); // 블로그 기본 정보 반환
     }
 
     @Override
-    public BlogDetailInfoResponse getDetailBlogInfo(CustomUserDetails userDetails, Optional<String> nickname) {
-        Long userId = nickname
-                .map(sharedService::getUserByNickname) // 닉네임으로 사용자 조회
-                .orElseGet(() -> sharedService.getUserByEmail(userDetails.getEmail()))  // 이메일로 사용자 조회
-                .getId(); // 사용자 ID 추출
+    @Transactional(readOnly = true)
+    public BlogDetailInfoResponse getDetailBlogInfo(CustomUserDetails userDetails, String nickname) {
+        User user = userRepository.findByNickname(nickname) // 닉네임으로 사용자 조회
+                .map(UserEntity::convertToDomain)
+                .orElseThrow(
+                        () -> new RestApiException(CommonExceptionCode.NOT_FOUND_USER)
+                );
 
-        Blog blog = sharedService.getBlogById(userId);
-        List<YearlyPostDto> yearlyPostCounts = postService.getYearlyPostCounts(blog.getId()); // 연간 포스트 수 조회
+        if (user.getEmail().equals(userDetails.getEmail())) {
+            throw new RestApiException(CommonExceptionCode.FORBIDDEN);
+        }
+
+        // 블로그 조회 및 연간 포스트 수 조회
+        Blog blog = getBlogById(user.getId());
+        List<YearlyPostDto> yearlyPostCounts = postService.getYearlyPostCounts(blog.getId());
 
         return blog.convertToBlogDetailInfoResponse(yearlyPostCounts); // 블로그 상세 정보 반환
     }
